+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.15.6 (2017-11-20)
 
 - Relax restriction on maximum terminal size from 256x256 to 1024x1024 cells. 
+- Add `output.texture-filter` option (value is `linear` or `nearest`).
 
 ### 0.15.5 (2017-11-13)